Prevent overwriting temp file that already exists.
This fixes an interaction with web servers that offer an automatic code reload
option, eg. paster or gunicorn. These temp files would be written each time
the server started, causing it to restart again in a loop.

+++ src/hypothesis/internal/reflection.py
@@ -365,22 +365,21 @@ def source_exec_as_module(source):
         )
         temporary_filepath = os.path.join(d, temporary_name + u'.py')
         final_filepath = os.path.join(d, final_name + u'.py')
-        f = open(temporary_filepath, u'w')
-        f.write(source)
-        f.close()
-        assert os.path.exists(temporary_filepath)
-
-        try:
-            os.rename(temporary_filepath, final_filepath)
-        except OSError:  # pragma: no cover
-            # The odds of final_filepath being a directory are basically zero,
-            # and it's basically impossible for them to be on different
-            # filesystems, so
-            # if this is raised it's because the destination already exists on
-            # Windows. That's fine, it won't be different, so just keep going,
-            # deleting our tempfile.
-            assert not os.path.isdir(final_filepath)
-            os.remove(temporary_filepath)
+        if not os.path.exists(final_filepath):
+            with open(temporary_filepath, u'w') as f:
+                f.write(source)
+            assert os.path.exists(temporary_filepath)
+            try:
+                os.rename(temporary_filepath, final_filepath)
+            except OSError:  # pragma: no cover
+                # The odds of final_filepath being a directory are basically zero,
+                # and it's basically impossible for them to be on different
+                # filesystems, so
+                # if this is raised it's because the destination already exists on
+                # Windows. That's fine, it won't be different, so just keep going,
+                # deleting our tempfile.
+                assert not os.path.isdir(final_filepath)
+                os.remove(temporary_filepath)
 
         assert os.path.exists(final_filepath)
         assert not os.path.exists(temporary_filepath)
